+++ packages/dnb-ui-lib/src/components/index.js
@@ -23,7 +23,7 @@ import LineTitle from './line-title/LineTitle'
 import Logo from './logo/Logo'
 import Modal from './modal/Modal'
 import Notification from './notification/Notification'
-import Progress from './progress/Progress'
+import ProgressIndicator from './progress-indicator/ProgressIndicator'
 import Slider from './slider/Slider'
 import StepIndicator from './step-indicator/StepIndicator'
 import Switch from './switch/Switch'
@@ -44,7 +44,7 @@ export {
   Logo,
   Modal,
   Notification,
-  Progress,
+  ProgressIndicator,
   Slider,
   StepIndicator,
   Switch,