+++ packages/dnb-ui-lib/src/components/lib.js
@@ -25,7 +25,7 @@ import LineTitle from './line-title/LineTitle'
 import Logo from './logo/Logo'
 import Modal from './modal/Modal'
 import Notification from './notification/Notification'
-import Progress from './progress/Progress'
+import ProgressIndicator from './progress-indicator/ProgressIndicator'
 import Slider from './slider/Slider'
 import StepIndicator from './step-indicator/StepIndicator'
 import Switch from './switch/Switch'
@@ -46,7 +46,7 @@ export {
   Logo,
   Modal,
   Notification,
-  Progress,
+  ProgressIndicator,
   Slider,
   StepIndicator,
   Switch,
@@ -68,7 +68,7 @@ export const getComponents = () => {
     Logo,
     Modal,
     Notification,
-    Progress,
+    ProgressIndicator,
     Slider,
     StepIndicator,
     Switch,