+++ packages/dnb-ui-lib/src/components/progress/index.js
@@ -1,8 +1,0 @@
-/**
- * Component Entry
- *
- */
-
-import Progress from './Progress'
-export default Progress
-export * from './Progress'