+++ packages/dnb-ui-lib/src/components/progress/style/themes/ui.js
@@ -1,6 +1,0 @@
-/**
- * Imports the default theme
- *
- */
-
-import './dnb-progress-theme-ui.scss'