+++ packages/dnb-ui-lib/src/index.js
@@ -23,7 +23,7 @@ import LineTitle from './components/line-title/LineTitle'
 import Logo from './components/logo/Logo'
 import Modal from './components/modal/Modal'
 import Notification from './components/notification/Notification'
-import Progress from './components/progress/Progress'
+import ProgressIndicator from './components/progress-indicator/ProgressIndicator'
 import Slider from './components/slider/Slider'
 import StepIndicator from './components/step-indicator/StepIndicator'
 import Switch from './components/switch/Switch'
@@ -64,7 +64,7 @@ export {
   Logo,
   Modal,
   Notification,
-  Progress,
+  ProgressIndicator,
   Slider,
   StepIndicator,
   Switch,